% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{construct_query}
\alias{construct_query}
\title{construct_query}
\usage{
construct_query(apikey, start, end, station_id, params, type)
}
\arguments{
\item{apikey}{An optional fmi-apikey.
See \url{http://en.ilmatieteenlaitos.fi/open-data-manual}.}

\item{start}{Start time as a POSIXt-object obtained with e.g.
lubridate::as_datetime().}

\item{end}{End time as a POSIXt-object. Defaults to \code{start}.}

\item{station_id}{The weather station id. Defaults to \code{100971}.
(Kaisaniemi weather station). A list of the available stations can be
downloaded with the \code{\link[=get_stations]{get_stations()}}-function.}

\item{params}{Query parameters, a comma separated string, such as
\code{"tday,rrday"}. For more options, see Details.}

\item{type}{Type of the query, must be one of the following:
\verb{"weather-daily", "weather-hourly", "airquality", "radiation"}.}
}
\value{
A list queries; one for each time chunk
}
\description{
Constructs the queries to get the data. Most importantly, splits it into
small enough chunks for the API with \code{\link[=mk_time_seq]{mk_time_seq()}}.
}
