% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_download.R
\name{fmi_airquality}
\alias{fmi_airquality}
\title{fmi_airquality}
\usage{
fmi_airquality(
  start,
  end = start,
  station_id = "100742",
  params = "aqindex_pt1h_avg",
  simplify_names = TRUE,
  verbose = FALSE,
  fmi_apikey = NA_character_
)
}
\arguments{
\item{start}{Start time as a POSIXt-object obtained with e.g.
lubridate::as_datetime().}

\item{end}{End time as a POSIXt-object. Defaults to \code{start}.}

\item{station_id}{The weather station id. Defaults to \code{100742}.
(Helsinki Mannerheimintie air quality station). A list of the available
stations can be downloaded with the \code{\link[=get_stations]{get_stations()}}-function.}

\item{params}{Query parameters, a comma separated string, such as
\code{"pm10_pt1h_avg,pm25_pt1h_avg"}. For more options, see Details.}

\item{simplify_names}{If \code{TRUE}, variable names are simplified (eg. \code{aqindex_pt1h_avg} is
converted to \code{airquality}).}

\item{verbose}{If \code{TRUE}, prints the progress. Defaults to \code{FALSE}.}

\item{fmi_apikey}{An optional fmi-apikey.
See \url{http://en.ilmatieteenlaitos.fi/open-data-manual}.}
}
\value{
A tibble with time in the first
column and the variables specified in \code{params} in the other columns.
}
\description{
Download data from the fmi (urban) airquality API. If the request is larger
than the api allows, it is split into multiple parts. If this results to more
than 30 queries, a warning message is printed.
}
\details{
The available variables are listed below, more information can be found from
\url{https://en.ilmatieteenlaitos.fi/open-data-manual}.

The following variables are available:
\itemize{
\item \code{aqindex_pt1h_avg}: air quality index
\item \code{co_pt1h_avg}: carbon monoxide
\item \code{no2_pt1h_avg}: nitrogen dioxide
\item \code{no_pt1h_avg}: nitrogen monoxide
\item \code{o3_pt1h_avg}: ozone
\item \code{pm10_pt1h_avg}: particulate matter < 10 nm
\item \code{pm25_pt1h_avg}: particulate matter < 2.5 nm
\item \code{qbcpm25_pt1h_avg}: black carbon pm2.5
\item \code{so2_pt1h_avg}: sulphur dioxide
\item \code{trsc_pt1h_avg}: odorous sulphur compounds
}
}
