% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{fmi_get}
\alias{fmi_get}
\title{fmi_get}
\usage{
fmi_get(queries, simplify_names, verbose)
}
\arguments{
\item{queries}{List of queries (urls).}

\item{simplify_names}{if \code{TRUE}, the column names will be simplified with
\code{\link[=name_map]{name_map()}}}

\item{verbose}{if \code{TRUE}, progress between queries is printed.}
}
\value{
A list of http results
}
\description{
Downloads the data from FMI API and extracts it from the xml-result. The
queries are executed with \code{purrr::get_slowly}.
}
