% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{get_stations}
\usage{
get_stations()
}
\value{
A tibble with the station name, id, and other information.
}
\description{
Download the available stations from
\url{http://en.ilmatieteenlaitos.fi/open-data-manual}. See especially the
variable \code{groups} which specifies wheter the station provides weather or
urban air quality data.
}
