% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{mk_time_seq}
\alias{mk_time_seq}
\title{mk_time_seq}
\usage{
mk_time_seq(start, end, hourly)
}
\arguments{
\item{start}{Start time as a POSIXt-object obtained with e.g.
lubridate::as_datetime().}

\item{end}{End time as a POSIXt-object. Defaults to \code{start}.}

\item{hourly}{if \code{TRUE}, download hourly observations instead of daily.}
}
\value{
A list queries; one for each time chunk
}
\description{
Make a sequence of (start,end)-pairs that cover the data.
}
