% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{name_map}
\alias{name_map}
\title{name_map}
\usage{
name_map(names)
}
\arguments{
\item{names}{Names to be simplified.}
}
\value{
A character vector with the simplified names
}
\description{
Simplify (column) names to be more legible. Eg. \code{t2m} => \code{temp} or
\code{co_pt1h_avg} => \code{carbon_monoxide}.
}
