% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_content}
\alias{process_content}
\title{process_content}
\usage{
process_content(res)
}
\arguments{
\item{res}{HTTP-result containing the observations as XML.}
}
\value{
The observations formatted in a tibble.
}
\description{
Extract the relevant data from the XML-result.
}
