% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{report_errors}
\alias{report_errors}
\title{report_errors}
\usage{
report_errors(res_list)
}
\arguments{
\item{res_list}{list of http-results.}
}
\value{
Does not return anything but stops execution if one of them returned
with an error code.
}
\description{
Check the http results for error codes. If there is at least one, stop
execution and print the result.
}
