% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_code_map.R
\name{weather_code_map}
\alias{weather_code_map}
\title{weather_code_map}
\usage{
weather_code_map(weather_code)
}
\arguments{
\item{weather_code}{The original weather code from FMI API.}
}
\value{
A character vector with the simplified values.
}
\description{
Map weather codes (wawa, WMO code 4680) to simplified values. See codes
and the respective values at
\url{https://artefacts.ceda.ac.uk/badc_datadocs/surface/code.html}.
}
