% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmi_download.R
\name{fmi_download}
\alias{fmi_download}
\title{fmi_download}
\usage{
fmi_download(fmi_apikey, start, end = start, station_id = "100971",
  params = ifelse(hourly, "t2m,r_1h", "tday,rrday"), hourly = FALSE,
  simplify_names = TRUE)
}
\arguments{
\item{fmi_apikey}{Your personal fmi-apikey.
See \url{http://en.ilmatieteenlaitos.fi/open-data-manual}.}

\item{start}{Start time as a POSIXt-object obtained with e.g.
lubridate::as_datetime().}

\item{end}{End time as a POSIXt-object. Defaults to \code{start}.}

\item{station_id}{The weather station id. Defaults to \code{100971}
(Kaisaniemi weather station). See
\url{http://en.ilmatieteenlaitos.fi/observation-stations}.}

\item{params}{Query parameters. If not provided, temprature and
precipitation is downloaded. See
\url{http://en.ilmatieteenlaitos.fi/open-data-manual-fmi-wfs-services}.}

\item{hourly}{If \code{TRUE}, hourly data is downloaded. Otherwise the daily
data is downloaded.}

\item{simplify_names}{If \code{TRUE}, an attempt is made to simplify some of
the names of the variables. For example, t2m/tday is temp etc.}
}
\value{
A tibble with date (& time if hourly data is requested) in the first
  column and the variables specified in \code{params} in the other columns.
}
\description{
Download data from the fmi weather API.
}
