% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{get_stations}
\alias{get_stations}
\title{get_stations}
\usage{
get_stations()
}
\value{
A tibble with the station name, id, latitude and longitude.
}
\description{
Download the available weather stations from
 \url{http://en.ilmatieteenlaitos.fi/open-data-manual}.
}
